# openapi.yaml
openapi: 3.0.3
info:
  title: Midterm Task API
  version: 1.0.0

servers:
  - url: http://localhost:3000

paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Server is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /api/tasks:
    get:
      summary: Get all tasks
      responses:
        "200":
          description: Get all tasks in the server
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        "404":
          description: Task does not exist
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create Task
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Task'
        "400":
          description: Error with user input
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/ErrorResponse'


      requestBody:
        description: Item data to send
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Task"
  /api/tasks/{id}:
    summary: Find, update, or delete a task
    parameters:
      - in: query
        name: id
        schema:
          type: number
    get:
      responses:
        "200":
          description: "Get a task"
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Task'
        "404":
          description: Task does not exist
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/ErrorResponse'

    put:
      responses:
        "200":
          description: "Update an item"
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Task'
        "400":
          description: Error with user input
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Task does not exist
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        description: Task data to send
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
    delete:
      responses:
        "201":
          description: "Delete a task"
          content:
            application/json:
              schema:
                type: string
        "404":
          description: "Task does not exist"
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/ErrorResponse'



  #
  # Remember:
  # - Express routes use :id, such as /tasks/:id.
  # - OpenAPI paths use {id}, such as /tasks/{id}.
  #
  # You may define schemas under components/schemas if you want to avoid
  # repeating the same object structure many times.

components:
  schemas:
    Task:
      type: object
      properties:
        title:
          example: "clock"
          type: string
        name:
          example: "clock"
          type: string
        quantity:
          example: 6
          type: integer
          format: uint
      required:
        - title
        - course
        - completed
    TaskInput:
      type: object
      properties:
        title:
          example: "Start Lab 5"
          type: string
        completed: 
          example: true
          type: boolean
        course:
          example: "CS553"
          type: string
      required:
        - title
        - course
        - completed
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Task does not exist"
      required:
        - error